+++ CHANGELOG.md
@@ -231,6 +231,7 @@
 207. Added UnityEngine.Vector2 Support to MelonPreferences.
 208. Added UnityEngine.Vector3 Support to MelonPreferences.
 209. Added UnityEngine.Vector4 Support to MelonPreferences.
+210. Updated MelonPreferences UnityEngine.Color Support to Conform to 0 - 255 RGBA Values in File.
 
 ---